+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Features
 --------
   - RGB+W intensity control
   - more sensitive logging to avoid floods on DEBUG level
+  - websocket logging, visible in browser console or CLI client
 
 Fixes
 -----